Simplifie les classes de notification par webhook
